requirements.txt
fastapi
uvicorn
biopython

